Split observer sync regression test
